+++ .prettierrc.js
@@ -1,8 +1,0 @@
-module.exports = {
-  arrowParens: 'avoid',
-  jsxBracketSameLine: true,
-  singleQuote: true,
-  trailingComma: 'es5',
-  tabWidth: 2,
-  semi: false,
-};