+++ examples/http-and-file/index.ts
@@ -1,13 +1,9 @@
-import { http, retry, file } from 'rewait'
+import { file, http, retry } from 'rewait'
 import { createServerAndFile } from './util'
 
 console.log(`Waiting for http://localhost:3000 and temp.txt`)
 createServerAndFile()
 
-retry([
-  http('http://localhost:3000/'),
-  file('temp.txt'),
-]).then(() => {
+retry([http('http://localhost:3000/'), file('temp.txt')]).then(() => {
   console.log('Ready!')
 })
-