+++ examples/simple.js
@@ -3,8 +3,10 @@
 const { retry, http } = require('../src')
 
 // Poll a Web service until it responds
-retry(http('http://localhost:8080')).then(() => {
-  console.log('Service is ready')
-}).catch(err => {
-  console.log('Error: ' + err)
-})
+retry(http('http://localhost:8080'))
+  .then(() => {
+    console.log('Service is ready')
+  })
+  .catch(err => {
+    console.log(`Error: ${err}`)
+  })