+++ spec/udp.spec.ts
@@ -1,4 +1,4 @@
-import dgram from 'dgram'
+import dgram from 'node:dgram'
 import test from 'tape'
 import udp from '../src/udp'
 
@@ -17,16 +17,21 @@ test('socket() connects via UDP4', t => {
   const port = 43424
   const server = dgram.createSocket('udp4')
   server.bind(port, () => {
-    udp(port)().then(client => {
-      t.ok(client instanceof dgram.Socket, 'callback receives socket instance')
-      client.on('close', () => {
-        t.ok(true, 'client socket closed')
-        t.throws(() => client.close(), 'socket already closed')
+    udp(port)()
+      .then(client => {
+        t.ok(
+          client instanceof dgram.Socket,
+          'callback receives socket instance'
+        )
+        client.on('close', () => {
+          t.ok(true, 'client socket closed')
+          t.throws(() => client.close(), 'socket already closed')
+        })
+        server.close(() => {
+          t.ok(true, 'server socket closed')
+        })
       })
-      server.close(() => {
-        t.ok(true, 'server socket closed')
-      })
-    }).catch(err => t.notOk(err, 'should not error'))
+      .catch(err => t.notOk(err, 'should not error'))
   })
 })