+++ src/MultiError.ts
@@ -1,5 +1,8 @@
 export default class MultiError extends Error {
-  constructor(message: string | undefined, public errors: any[] = []) {
+  constructor(
+    message: string | undefined,
+    public errors: any[] = []
+  ) {
     super(message)
   }
 }