+++ src/sequence.ts
@@ -1,14 +1,17 @@
-import { CheckFunction } from './fn'
+import type { CheckFunction } from './fn'
+import type { CheckResults } from './retry'
 
 /**
  * Perform a set of checks in sequence instead of in parallel.
  */
-export default function sequence<T extends CheckFunction[]>(...fns: T) {
-  return async function () {
+export default function sequence<T extends readonly CheckFunction[]>(
+  ...fns: T
+) {
+  return async (): Promise<CheckResults<T>> => {
     const results = []
     for (let idx = 0; idx < fns.length; idx++) {
       results.push(await fns[idx]())
     }
-    return results
+    return results as CheckResults<T>
   }
 }