+++ src/udp.ts
@@ -1,4 +1,4 @@
-import dgram from 'dgram'
+import dgram from 'node:dgram'
 
 function getUdpType(host?: string) {
   if (!host || typeof host !== 'string') {
@@ -12,12 +12,15 @@ function getUdpType(host?: string) {
   }
 }
 
-
 /**
  * Promisify dgram.connect()
  */
-function dgramConnect(port: number, address: string | undefined, opts: dgram.SocketOptions) {
-  return new Promise<dgram.Socket>((resolve) => {
+function dgramConnect(
+  port: number,
+  address: string | undefined,
+  opts: dgram.SocketOptions
+) {
+  return new Promise<dgram.Socket>(resolve => {
     // TODO: use dns.lookup() to determine address family?
     const socket = dgram.createSocket(opts)
     socket.connect(port, address, () => {
@@ -27,15 +30,22 @@ function dgramConnect(port: number, address: string | undefined, opts: dgram.Soc
 }
 
 export interface CheckUdpOptions {
-  checkOk: (socket: dgram.Socket, opts: CheckUdpOptions) => any | Promise<any>
+  /**
+   * Throw an error to indicate a not-ok state. The return value is ignored.
+   */
+  checkOk: (socket: dgram.Socket, opts: CheckUdpOptions) => unknown
   close: boolean
   socketOptions: dgram.SocketOptions
 }
 
-export default function checkUdp(port: number, address?: string, userOpts?: Partial<CheckUdpOptions>) {
+export default function checkUdp(
+  port: number,
+  address?: string,
+  userOpts?: Partial<CheckUdpOptions>
+) {
   const opts: CheckUdpOptions = {
     close: true,
-    checkOk: () => true,
+    checkOk: () => {},
     socketOptions: {
       type: getUdpType(address) || 'udp4',
       ...userOpts?.socketOptions,
@@ -43,21 +53,18 @@ export default function checkUdp(port: number, address?: string, userOpts?: Part
     ...userOpts,
   }
   return async () => {
-    let client: dgram.Socket | undefined = undefined
+    let client: dgram.Socket | undefined
     let open: boolean = false
     try {
       client = await dgramConnect(port, address, opts.socketOptions)
       open = true
-      const ok = await opts.checkOk(client, opts)
-      if (!ok) {
-        throw new Error(`UDP connection to ${address}:${port} failed checkOk()`)
-      }
+      await opts.checkOk(client, opts)
       return client
     } finally {
       if (client && open && opts.close) {
         try {
           client.close()
-        } catch (err) {
+        } catch {
           // TODO how to handle this? we can't re-throw
           // ERR_SOCKET_DGRAM_NOT_RUNNING
           // console.error(err)